cart list update

+++ shoppingmall/src/main/java/com/shop/shoppingmall/controller/AdminController.java
@@ -8,7 +8,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Controller
@@ -22,16 +25,28 @@ public class AdminController {
     }
 
     @GetMapping
-    public String manageItem(Model model) {
+    public String manageItem(Model model, HttpSession session) {
+        if(!session.getAttribute("userId").equals("admin@admin")) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+
         List<Item> item = itemService.manageItems();
         model.addAttribute("item", item);
+        model.addAttribute("userId", session.getAttribute("userId"));
+        model.addAttribute("userName", session.getAttribute("userName"));
         return "admin/manageItem";
     }
 
     @GetMapping("/edit/{id}")
-    public String editItem(Model model, @PathVariable Long id) {
+    public String editItem(Model model, @PathVariable Long id, HttpSession session) {
+        if(!session.getAttribute("userId").equals("admin@admin")) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+
         Item item = itemService.findById(id);
-        model.addAttribute("item", new ItemEditDto(id, item.getCode(), item.getName(), item.getPrice(), item.getStock(), item.getStatus()));
+        model.addAttribute("item", new ItemEditDto(id, item.getItemCode(), item.getName(), item.getPrice(), item.getStock(), item.getStatus()));
+        model.addAttribute("userId", session.getAttribute("userId"));
+        model.addAttribute("userName", session.getAttribute("userName"));
 
         return "admin/editItem";
     }
@@ -43,19 +58,33 @@ public class AdminController {
     }
 
     @GetMapping("/add")
-    public String addItem(Model model) {
+    public String addItem(Model model, HttpSession session) {
+        if(!session.getAttribute("userId").equals("admin@admin")) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+
         model.addAttribute("item", new ItemAddDto("", null, 0, 0, null));
+        model.addAttribute("userId", session.getAttribute("userId"));
+        model.addAttribute("userName", session.getAttribute("userName"));
         return "admin/addItem";
     }
 
     @PostMapping("/add")
-    public String addItem(@ModelAttribute ItemAddDto add) {
-        itemService.itemAdd(add);
+    public String addItem(@ModelAttribute ItemAddDto add, HttpSession session, MultipartFile imgFile) {
+        if(!session.getAttribute("userId").equals("admin@admin")) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+
+        itemService.itemAdd(add, imgFile);
         return "redirect:/manage";
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteItem(@PathVariable Long id) {
+    public String deleteItem(@PathVariable Long id, HttpSession session) {
+        if(!session.getAttribute("userId").equals("admin@admin")) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+
         itemService.deleteItem(id);
         return "redirect:/manage";
     }
